--- src/main/java/com/kosa/kosafinalprojbaeken/chatting/config/WebSocketBrokerConfig.java
@@ -1,4 +1,0 @@
-package com.kosa.kosafinalprojbaeken.chatting.config;
-
-public class WebSocketBrokerConfig {
-}
