+++ src/main/java/com/kosa/kosafinalprojbaekend/chatting/config/WebSocketBrokerConfig.java
@@ -1,0 +1,4 @@
+package com.kosa.kosafinalprojbaekend.chatting.config;
+
+public class WebSocketBrokerConfig {
+}
